demo-two/requirements.txt
capiscio-sdk>=2.5.0
capiscio-mcp[mcp,crypto]>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
